Added Test Files in TestJournals.java
This file contains a main and many different functions that can be used to interface with the other modules for testing purposes. To use, on line 18, replace the function that is written there with a different function that you wish to test. In order to test different numbers, change the variables that monitor the loop counters to larger or smaller numbers to fit needs.

This file has functions to test
-creating a journal time
-changing journal title time
-changing journal cover colour time
-delete journals time
-open journals time
-create pages time

Two new functions added within Journal.java to circumvent user input when creating hundreds of journals/pages at a time. These functions are createPagesForTest() and openJournalForTest()